edge cases in facade/service layer for order fixed

+++ service/src/main/java/cz/fi/muni/pa165/service/facade/OrderFacadeImpl.java
@@ -35,7 +35,15 @@ public class OrderFacadeImpl implements OrderFacade {
 
     @Override
     public Long createOrder(OrderCreateDTO order, String userLogin) {
-
+        if (order == null) {
+            throw new IllegalArgumentException("Order cannot be null.");
+        }
+        if (order.getServices().isEmpty() && order.getTires().isEmpty()) {
+            throw new IllegalArgumentException("Cannot create empty order.");
+        }
+        if (userLogin.isEmpty()) {
+            throw new IllegalArgumentException("Cannot assign order to empty user.");
+        }
         Order newOrder = new Order();
         newOrder.setState(order.getState());
         newOrder.setDateOfOrder(order.getDateOfOrder());
@@ -45,11 +53,23 @@ public class OrderFacadeImpl implements OrderFacade {
 
     @Override
     public void addTireToOrder(Long orderId, Long tireId) {
+        if (orderId == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        if (tireId == null) {
+            throw new IllegalArgumentException("Tire id cannot be null");
+        }
         orderService.addTireToOrder(orderId, tireId);
     }
 
     @Override
     public void addServiceToOrder(Long orderId, Long serviceId) {
+        if (orderId == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        if (serviceId == null) {
+            throw new IllegalArgumentException("Service id cannot be null");
+        }
         orderService.addServiceToOrder(orderId, serviceId);
     }
 
@@ -82,20 +102,34 @@ public class OrderFacadeImpl implements OrderFacade {
 
     @Override
     public void finishOrder(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         orderService.finish(orderService.findById(id));
     }
 
     @Override
     public void cancelOrder(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         orderService.cancel(orderService.findById(id));
     }
 
     @Override
     public void removeOrder(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         orderService.remove(orderService.findById(id));
     }
 
     @Override
-    public void confirmOrder(Long id) {orderService.confirm(orderService.findById(id)); }
+    public void confirmOrder(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        orderService.confirm(orderService.findById(id));
+    }
 
 }
